Danielbot is a state machine + other goodies
Hey guys I made a change in bot.js.  There is a new function called
“getRandom()”  It should take in two numbers as arguments and return a
random number between the two values given.  The first value in the
arguments is the lower value and the second is the upper.  I gave an
example in the comments of how to use it.  Let me know if it doesn’t
work or you get some weird numbers being spit out by it.