Enabled Arcade Physics & stuck bots in a group
I'm surprised that this game had been running so well without having
physics enabled. Now that this is enabled, hopefully we could now do
some of the stuff we could find here
(http://phaser.io/examples/v2/category/arcade-physics) without getting
errors.

I've also made a phaser group that all out bots belong in. This does
nothing so far, but this opens the door for us to add things or events
that affect all of our bots at once.

Feel free to revert this if this is unwanted