Cleaning up code
I figure the best way to build an idea is to start with a solid
foundation of motion and emotion states. The code for states is very
similar to the jeff,js file but it is my hope that I will be able to
branch off from it and add my own stuff once I get a better
understanding of ow to implement it.