Total bot refactor
For ease of use, I have completely refactored my bot’s code to mirror
jeff.js bot. I have kept my old code in a separate file to go back and
work but for the time being my bot will be like this since this is the
safest option. expect a completed and original code for faust.js by the
end of spring break.