Goals and Memory stuff
Added some preliminary goals and productions based around them.  It is
all working but I will continue working on some stuff including but not
limited to making it so that only one of the boredom related goals can
fires at a time, overriding all movement except for production movement
until a goal is met, making more random boredom goals, and more
responsive removal of goals when their conditions are met.